Refactor: encapuslate colony home layout actions panel
