+++ src/components/common/ColonyHome/ColonyHomeLayout/ColonyHomeLayout.css
