+++ src/components/common/ColonyHome/ColonyHomeLayout/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ColonyHomeLayout';
